# Improve performance of ConvertSpanUniquenessMetrics