<!-- Checklist For PRs

* Ensure you've written tests where applicable
* Add a CHANGELOG.md entry describing your change, thank yourself!
* Document any changes to metrics or configuration
-->


#### Summary
<!-- Simple summary of what the code does or what you have changed. -->

Improve performance of ConvertSpanUniquenessMetrics, which is called (albeit asynchronously) in the hot path of our `metric_extraction` sink ingestion during span processing.

Before:

```
BenchmarkConvertSpanUniquenessMetrics-8           500000              2222 ns/op             960 B/op         21 allocs/op
BenchmarkParseMetricSSF-8                        1000000              1786 ns/op             440 B/op         19 allocs/op
```

After:

```
BenchmarkConvertSpanUniquenessMetrics-8          1000000              1536 ns/op             800 B/op         13 allocs/op
BenchmarkParseMetricSSF-8                        1000000              1053 ns/op             264 B/op          9 allocs/op
```

#### Motivation
<!-- Why are you making this change? -->


#### Test plan
<!-- How did you test this change? This can be as simple as “I wrote automated tests.” -->


#### Rollout/monitoring/revert plan
<!-- Instructions for deploying, monitoring, and reverting this change. -->

r? @stripe/observability